Drawer를 기본으로 각 Component를 사용하는 예제 완성
- Main.java 클래스 실행시 확인 가능
-> README.md로 정리 필요

+++ src/main/java/jy/practice/dto/RoundDto.java
@@ -26,7 +26,7 @@ public class RoundDto {
   @JsonProperty("end_date") // yyyy-MM-dd
   private String endDate;
   @JsonProperty("achievement_rate")
-  private Double achievementRate;
+  private Integer achievementRate;
   @JsonProperty("status")
   private Status status;
   @JsonProperty("created_date") // yyyy-MM-dd
